requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
aws-xray-sdk==0.95
boto==2.48.0
boto3==1.7.31
botocore==1.10.31
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cookies==2.2.1
coverage==4.5.1
cryptography==2.2.2
docker==3.3.0
docker-pycreds==0.3.0
docutils==0.14
Flask==1.0.2
Flask-CAS==1.0.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.1
jsonpickle==0.9.6
MarkupSafe==1.0
mock==2.0.0
more-itertools==4.2.0
moto==1.3.3
pbr==4.0.4
pluggy==0.6.0
py==1.5.3
pyaml==17.12.1
pycparser==2.18
pytest==3.6.1
pytest-cov==2.5.1
python-dateutil==2.6.1
python-dotenv==0.8.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-mock==1.5.0
responses==0.9.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
websocket-client==0.48.0
Werkzeug==0.14.1
wrapt==1.10.11
xmltodict==0.11.0
